uclamp_test_thermal_pressure: Enhance one warning
Detecting if we are running on smallest fitting uclamp_min was noisy.
Improve the logic so that it's better now. It still can fail if
uclamp_min happens to be  more energy efficient than the smallest cpu.

It's more work than we can do now to detect if it's the most energy
efficient. Spill out the warning with a suggestion for the user to
check.

Signed-off-by: Qais Yousef <qyousef@layalina.io>